technical_demo/requirements.txt
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
sqlite3